made changes as per review 2.3.
